backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ev_recharge"
version = "0.1.0"
dependencies = [
    "flask",
    "flask-cors",
    "mysql-connector-python",
    "python-dotenv",
    "PyJWT"
]